simplify prettier command execution
